requirements.txt
asgiref==3.3.4
backcall==0.2.0
colorama==0.4.4
decorator==5.0.9
Django==3.2.4
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2==2.8.6
Pygments==2.9.0
python-dotenv==0.20.0
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
